refactor: use type alias to import
