+++ front-end/src/App.vue
@@ -1,13 +1,13 @@
 <script setup lang="ts">
-import { RouterView } from "vue-router";
-import NavBar from "./components/NavBar.vue";
-import Header from "./components/Header.vue";
-import PopupRegister from "./components/PopupRegister.vue";
 import { onMounted } from "vue";
+import { RouterView } from "vue-router";
 import type { CallbackTypes } from "vue3-google-login";
+import NavBar from "@/components/NavBar.vue";
+import Header from "@/components/Header.vue";
+import PopupRegister from "@/components/PopupRegister.vue";
 import { logout } from "@/helpers/logout";
 import { login } from "@/helpers/login";
-import { userStore } from "./store";
+import { userStore } from "@/store";
 
 onMounted(() => {
   const credential = localStorage.getItem("credential");
