+++ front-end/src/main.ts
@@ -1,8 +1,9 @@
-import "./assets/main.css";
+import "@/assets/main.css";
 import { createApp } from "vue";
-import App from "./App.vue";
-import router from "./router";
 import vue3GoogleLogin from "vue3-google-login";
+import App from "@/App.vue";
+import router from "@/router";
+
 const app = createApp(App);
 app.use(vue3GoogleLogin, {
   clientId: import.meta.env.VITE_GOOGLE_CLIENT_ID,
