+++ front-end/src/types/IUserStore.ts
@@ -1,0 +1,8 @@
+export interface IUserStore {
+  id?: string;
+  loggedIn?: boolean;
+  name?: string;
+  email?: string;
+  profilePicture?: string;
+  registerDone?: boolean;
+}
